Move requests and responses to a jackson folder to show dependency
